Changes requested by rsummers
